+++ binary_choice_game/views/QnPage.html
