+++ binary_choice_game/views/StartPage.html
@@ -1,0 +1,3 @@
+{{ block content }}
+  {{ rec_desc }}
+{{ endblock }}
